Persist language option changes via config file
